+++ src/simulake/simulake.hpp
@@ -8,7 +8,6 @@
  */
 
 #include <glad/glad.h>
-
 #include <GLFW/glfw3.h>
 
 #ifdef __APPLE__
